[ESI] Make MMIO data 64-bit
Since MMIO will be used to transmit things like pointers, we should
probably make it 64-bit data. This seemingly simple change had a bunch
of ripple effects.

+++ lib/Dialect/ESI/runtime/cpp/include/esi/Services.h
@@ -93,8 +93,8 @@ public:
 class MMIO : public Service {
 public:
   virtual ~MMIO() = default;
-  virtual uint32_t read(uint32_t addr) const = 0;
-  virtual void write(uint32_t addr, uint32_t data) = 0;
+  virtual uint64_t read(uint32_t addr) const = 0;
+  virtual void write(uint32_t addr, uint64_t data) = 0;
   virtual std::string getServiceSymbol() const override;
 };
